% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-clusters.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Calculate K-means clusters}
\usage{
calculate_clusters(data, k = 3)
}
\arguments{
\item{data}{A numeric data frame or matrix}

\item{k}{The number of clusters (default = 3)}
}
\value{
A K-means object containing cluster assignments and centers
}
\description{
This function runs K-means clustering on a numeric dataset.
}
