% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_functions .R
\name{my_kmeans}
\alias{my_kmeans}
\title{Compute k-means clusters}
\usage{
my_kmeans(x, k)
}
\arguments{
\item{x}{Numeric matrix or data frame to cluster}

\item{k}{Number of clusters}
}
\value{
A list containing cluster assignments and centroids
}
\description{
Compute k-means clusters
}
