% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot K-means clusters}
\usage{
plot_clusters(data, clusters)
}
\arguments{
\item{data}{Numeric dataframe used for clustering}

\item{clusters}{K-means result object from calculate_clusters()}
}
\description{
Plots a 2D scatter plot of the data with points colored by cluster.
Cluster centers are highlighted with a larger symbol.
}
