% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions .R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Plot k-means clusters (2D)}
\usage{
plot_kmeans(x, clusters)
}
\arguments{
\item{x}{Numeric matrix with 2 columns}

\item{clusters}{Vector of cluster assignments}
}
\value{
ggplot2 plot of clusters
}
\description{
Plot k-means clusters (2D)
}
\examples{
set.seed(123)
data <- matrix(rnorm(20), ncol = 2)
km <- my_kmeans(data, 3)
plot_kmeans(data, km$cluster)
}
