% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runkmeansApp.R
\name{runKmeansApp}
\alias{runKmeansApp}
\title{Launch the K-means Shiny App}
\usage{
runKmeansApp()
}
\description{
This function launches the Shiny app included in the package.
The app allows the user to select a dataset, set the number of clusters,
and view the clustering results as a plot and table.
}
